testing/mochitest/tests/test_bug351601.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=351601
-->
<head>
  <title>Test for Bug 351601</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=351601">Mozilla Bug 351601</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 351601 **/

var foo;
var isOK = false;
try {
  foo = navigator.buildID;
  isOK = true;
} catch (ex) { /* do not do anything here, as we are checking result below. */ }

if (foo == "0000000000")
  ok(isOK, "navigator.buildID should never throw");
else
  ok(isOK, "navigator.buildID is not 0000000000, but we can confirm the normal case does not crash.");
</script>
</pre>
</body>
</html>
